Firepunch and most status code
Firepunch should be working, I just need to test it a little more to make sure it burns as expected.  Ice and Thunder punch are working mostly, and I just need to make sure they work.  Damage taken from burn, poison, etc, and paralize, sleep, frozen need to still be coded.  This will be done as they come up but the foundation is in place.

Status also was changed into a class instead of a dictionary.